[!] Possible fix/hack for SM64 stars blend [!] Calling proper function for clearing Zbuffer(also adding proper cast for passed variable)
git-svn-id: https://subversion.assembla.com/svn/Daedalusx64/trunk@129 42e9bfbe-799a-4a2d-bad1-236e862a387a

+++ Source/Graphics/GraphicsContext.h
@@ -51,7 +51,7 @@ public:
 	virtual void SwitchToLcdDisplay() = 0;
 
 	virtual void ClearAllSurfaces() = 0;
-	virtual void ClearZBuffer(float depth) = 0;
+	virtual void ClearZBuffer(u32 depth) = 0;
 	virtual void Clear(bool clear_screen, bool clear_depth) = 0;
 	virtual void Clear(u32 frame_buffer_col, u32 depth) = 0;
 	virtual	void BeginFrame() = 0;
